refactoring core folder
