+++ backend/src/domain/core/formatters/discord_formatter.ts
