+++ backend/src/domain/core/formatters/string_builder.ts
