+++ backend/src/infrastructure/database/postgres/postgres.ts
@@ -1,5 +1,5 @@
+import { getEnv } from '@domain/core/utils/env';
 import { Pool, PoolClient } from 'pg';
-import { getEnv } from 'src/core/utils/env';
 import { v4 as uuidv4 } from 'uuid';
 
 let pool: PostgresPool | undefined = undefined;
